+++ src/Console/Command/Docker/Mysql/GenerateMetadata.php
@@ -106,7 +106,8 @@ class GenerateMetadata extends \Symfony\Component\Console\Command\Command
     {
         $containerName = $input->getArgument(self::COMMAND_ARGUMENT_CONTAINER);
         $mysqlService = $this->mysql->initialize($containerName);
-        $containerMetadata = $this->dockerContainer->inspectJsonWithDecode($containerName)[0];
+        $inspectResult = $this->dockerContainer->inspectJsonWithDecode($containerName);
+        $containerMetadata = $inspectResult[0] ?? throw new \RuntimeException("No inspect data for container '$containerName'");
         $vendorImage = $this->getVendorImageFromEnv($mysqlService, $containerMetadata);
         $output->writeln("Detected DB image: <info>$vendorImage</info>");