+++ src/Filesystem/Filesystem.php
@@ -188,7 +188,7 @@ class Filesystem extends \Symfony\Component\Filesystem\Filesystem implements Pro
     /**
      * Must move elsewhere in case new methods are added
      *
-     * @return array<string, string|array<string, string>>
+     * @return array<string, mixed>
      * @throws \JsonException
      */
     public function getAuthJson(): array
@@ -218,7 +218,7 @@ class Filesystem extends \Symfony\Component\Filesystem\Filesystem implements Pro
         }
 
         $allowedPaths = [
-            realpath(sys_get_temp_dir()),
+            realpath(sys_get_temp_dir()) ?: sys_get_temp_dir(),
             $this->getHostsFilePath(),
             $this->env->getProjectsRootDir()
         ];
@@ -232,7 +232,7 @@ class Filesystem extends \Symfony\Component\Filesystem\Filesystem implements Pro
 
         foreach ($paths as $path) {
             if ($this->exists($path)) {
-                $path = realpath($path);
+                $path = realpath($path) ?: $path;
             }
 
             $isAllowed = false;